requirements.txt
gspread==6.1.4
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.154.0
selenium==4.27.1
webdriver-manager==4.0.2
python-dotenv==1.0.1
stripe>=8.0.0
